fix size calculation again

+++ database/engine/pagecache.c
@@ -664,7 +664,7 @@ void init_page_cache(void)
 
         main_mrg = mrg_create();
 
-        size_t target_cache_size = default_rrdeng_page_cache_mb * 1024 * 1024;
+        size_t target_cache_size = (size_t)default_rrdeng_page_cache_mb * 1024ULL * 1024ULL;
         size_t main_cache_size = (target_cache_size / 100) * 95;
         size_t open_cache_size = 0;
         size_t extent_cache_size = (target_cache_size / 100) * 5;
